+++ package.json
@@ -17,6 +17,7 @@
     "react": "18.0.0",
     "react-native": "0.69.1",
     "react-native-config": "^1.4.6",
+    "react-native-css-vh-vw": "^1.0.5",
     "react-native-curved-bottom-bar": "^2.0.1",
     "react-native-nmap": "^0.0.66",
     "react-native-safe-area-context": "^4.3.1",
@@ -24,7 +25,8 @@
     "react-native-svg": "^12.3.0",
     "react-native-utils-scale": "^2.6.0",
     "react-native-vector-icons": "^9.2.0",
-    "react-native-vector-image": "^0.3.3"
+    "react-native-vector-image": "^0.3.3",
+    "recoil": "^0.7.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
